bypass the lock related with Transaction
